+++ instrumentation/opentelemetry-instrumentation-django/tests/test_middleware.py
@@ -22,7 +22,11 @@ from unittest.mock import Mock, patch
 from django import VERSION, conf
 from django.http import HttpRequest, HttpResponse
 from django.test.client import Client
-from django.test.utils import setup_test_environment, teardown_test_environment
+from django.test.utils import (
+    override_settings,
+    setup_test_environment,
+    teardown_test_environment,
+)
 
 from opentelemetry import trace
 from opentelemetry.instrumentation._semconv import (
@@ -33,7 +37,7 @@ from opentelemetry.instrumentation._semconv import (
 )
 from opentelemetry.instrumentation.django import (
     DjangoInstrumentor,
-    _DjangoMiddleware,
+    DjangoMiddleware,
 )
 from opentelemetry.instrumentation.propagators import (
     TraceResponsePropagator,
@@ -100,6 +104,18 @@ urlpatterns = [
 _django_instrumentor = DjangoInstrumentor()
 
 
+class CustomMiddleware(DjangoMiddleware):
+    pass
+
+
+class RandomMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        return self.get_response(request)
+
+
 # pylint: disable=too-many-public-methods
 class TestMiddleware(WsgiTestBase):
     @classmethod
@@ -136,11 +152,11 @@ class TestMiddleware(WsgiTestBase):
         self.env_patch.start()
         _django_instrumentor.instrument()
         self.exclude_patch = patch(
-            "opentelemetry.instrumentation.django.middleware.otel_middleware._DjangoMiddleware._excluded_urls",
+            "opentelemetry.instrumentation.django.middleware.otel_middleware.DjangoMiddleware._excluded_urls",
             get_excluded_urls("DJANGO"),
         )
         self.traced_patch = patch(
-            "opentelemetry.instrumentation.django.middleware.otel_middleware._DjangoMiddleware._traced_request_attrs",
+            "opentelemetry.instrumentation.django.middleware.otel_middleware.DjangoMiddleware._traced_request_attrs",
             get_traced_request_attrs("DJANGO"),
         )
         self.exclude_patch.start()
@@ -166,9 +182,9 @@ class TestMiddleware(WsgiTestBase):
         else:
             middleware = conf.settings.MIDDLEWARE_CLASSES
         # adding two dummy middlewares
-        temprory_middelware = "django.utils.deprecation.MiddlewareMixin"
-        middleware.append(temprory_middelware)
-        middleware.append(temprory_middelware)
+        temporary_middleware = "django.utils.deprecation.MiddlewareMixin"
+        middleware.append(temporary_middleware)
+        middleware.append(temporary_middleware)
 
         middleware_position = 1
         _django_instrumentor.instrument(
@@ -176,7 +192,83 @@ class TestMiddleware(WsgiTestBase):
         )
         self.assertEqual(
             middleware[middleware_position],
-            "opentelemetry.instrumentation.django.middleware.otel_middleware._DjangoMiddleware",
+            "opentelemetry.instrumentation.django.middleware.otel_middleware.DjangoMiddleware",
+        )
+
+    def test_custom_middleware_added_at_position(self):
+        _django_instrumentor.uninstrument()
+        if DJANGO_2_0:
+            middleware = conf.settings.MIDDLEWARE
+        else:
+            middleware = conf.settings.MIDDLEWARE_CLASSES
+        # adding two dummy middlewares
+        temporary_middleware = "django.utils.deprecation.MiddlewareMixin"
+        middleware.append(temporary_middleware)
+        middleware.append(temporary_middleware)
+        custom_middleware = (
+            CustomMiddleware.__module__ + "." + CustomMiddleware.__qualname__
+        )
+        middleware_position = 1
+
+        with override_settings(OTEL_DJANGO_MIDDLEWARE=custom_middleware):
+            _django_instrumentor.instrument(
+                middleware_position=middleware_position
+            )
+
+            self.assertEqual(
+                middleware[middleware_position], custom_middleware
+            )
+
+            _django_instrumentor.uninstrument()  # Uninstrument with the settings overridden
+
+    def test_random_middleware_refused_in_favor_of_django_middleware(self):
+        _django_instrumentor.uninstrument()
+        if DJANGO_2_0:
+            middleware = conf.settings.MIDDLEWARE
+        else:
+            middleware = conf.settings.MIDDLEWARE_CLASSES
+        # adding two dummy middlewares
+        temporary_middleware = "django.utils.deprecation.MiddlewareMixin"
+        middleware.append(temporary_middleware)
+        middleware.append(temporary_middleware)
+        random_middleware = (
+            RandomMiddleware.__module__ + "." + RandomMiddleware.__qualname__
+        )
+        middleware_position = 1
+
+        with override_settings(OTEL_DJANGO_MIDDLEWARE=random_middleware):
+            _django_instrumentor.instrument(
+                middleware_position=middleware_position
+            )
+
+        self.assertEqual(
+            middleware[middleware_position],
+            "opentelemetry.instrumentation.django.middleware.otel_middleware.DjangoMiddleware",
+        )
+
+    def test_custom_middleware_does_not_exist_default_to_django_middleware(
+        self,
+    ):
+        _django_instrumentor.uninstrument()
+        if DJANGO_2_0:
+            middleware = conf.settings.MIDDLEWARE
+        else:
+            middleware = conf.settings.MIDDLEWARE_CLASSES
+        # adding two dummy middlewares
+        temporary_middleware = "django.utils.deprecation.MiddlewareMixin"
+        middleware.append(temporary_middleware)
+        middleware.append(temporary_middleware)
+        non_existing_middleware = "path.to.a.non.existing.Middleware"
+        middleware_position = 1
+
+        with override_settings(OTEL_DJANGO_MIDDLEWARE=non_existing_middleware):
+            _django_instrumentor.instrument(
+                middleware_position=middleware_position
+            )
+
+        self.assertEqual(
+            middleware[middleware_position],
+            "opentelemetry.instrumentation.django.middleware.otel_middleware.DjangoMiddleware",
         )
 
     def test_middleware_added_at_position_if_wrong_position(self):
@@ -186,8 +278,8 @@ class TestMiddleware(WsgiTestBase):
         else:
             middleware = conf.settings.MIDDLEWARE_CLASSES
         # adding middleware
-        temprory_middelware = "django.utils.deprecation.MiddlewareMixin"
-        middleware.append(temprory_middelware)
+        temporary_middleware = "django.utils.deprecation.MiddlewareMixin"
+        middleware.append(temporary_middleware)
         middleware_position = (
             756  # wrong position out of bound of middleware length
         )
@@ -196,7 +288,7 @@ class TestMiddleware(WsgiTestBase):
         )
         self.assertEqual(
             middleware[0],
-            "opentelemetry.instrumentation.django.middleware.otel_middleware._DjangoMiddleware",
+            "opentelemetry.instrumentation.django.middleware.otel_middleware.DjangoMiddleware",
         )
 
     def test_templated_route_get(self):
@@ -605,12 +697,12 @@ class TestMiddleware(WsgiTestBase):
             response_hook_args = (span, request, response)
             response["hook-header"] = "set by hook"
 
-        _DjangoMiddleware._otel_request_hook = request_hook
-        _DjangoMiddleware._otel_response_hook = response_hook
+        DjangoMiddleware._otel_request_hook = request_hook
+        DjangoMiddleware._otel_response_hook = response_hook
 
         response = Client().get("/span_name/1234/")
-        _DjangoMiddleware._otel_request_hook = (
-            _DjangoMiddleware._otel_response_hook
+        DjangoMiddleware._otel_request_hook = (
+            DjangoMiddleware._otel_response_hook
         ) = None
 
         self.assertEqual(response["hook-header"], "set by hook")
@@ -636,9 +728,9 @@ class TestMiddleware(WsgiTestBase):
             # pylint: disable=broad-exception-raised
             raise Exception("request hook exception")
 
-        _DjangoMiddleware._otel_request_hook = request_hook
+        DjangoMiddleware._otel_request_hook = request_hook
         Client().get("/span_name/1234/")
-        _DjangoMiddleware._otel_request_hook = None
+        DjangoMiddleware._otel_request_hook = None
 
         # ensure that span ended
         finished_spans = self.memory_exporter.get_finished_spans()
@@ -649,9 +741,9 @@ class TestMiddleware(WsgiTestBase):
             # pylint: disable=broad-exception-raised
             raise Exception("response hook exception")
 
-        _DjangoMiddleware._otel_response_hook = response_hook
+        DjangoMiddleware._otel_response_hook = response_hook
         Client().get("/span_name/1234/")
-        _DjangoMiddleware._otel_response_hook = None
+        DjangoMiddleware._otel_response_hook = None
 
         # ensure that span ended
         finished_spans = self.memory_exporter.get_finished_spans()